Fixed exception after quit
This bug was silly.  It caused an exception when python was
cleaning up twisted before closing.  It seems there was a callback
in the stack that was trying to be applied to an object already
released.  This happened only when there were callbacks on
callbacks in twisted.

It seems that modules imported insude functions or methods
are cleaned up differently than when they are imported at the
beginning of a module.  Or maybe is a bug in either twisted or
bsddb3 that mess the garbage collector up.

https://bugzilla.gnome.org/show_bug.cgi?id=535572

Signed-off-by: Germán Poo-Caamaño <gpoo@gnome.org>